requirements.txt
aistudio-sdk==0.3.8
annotated-types==0.7.0
anyio==4.11.0
bce-python-sdk==0.9.53
black==25.11.0
certifi==2025.11.12
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
colorlog==6.10.1
easyocr==1.7.2
filelock==3.20.0
fsspec==2025.10.0
future==1.0.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.1.5
idna==3.11
ImageIO==2.37.2
imagesize==1.4.1
Jinja2==3.1.6
lazy_loader==0.4
MarkupSafe==3.0.3
modelscope==1.32.0
mpmath==1.3.0
mypy_extensions==1.1.0
Naked==0.1.32
networkx==3.5
ninja==1.13.0
numpy==2.2.6
opencv-contrib-python==4.10.0.84
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
opt-einsum==3.3.0
packaging==25.0
paddleocr==3.3.2
paddlepaddle==3.2.2
paddlex==3.3.9
pandas==2.3.3
pathspec==0.12.1
pillow==12.0.0
platformdirs==4.5.0
prettytable==3.17.0
protobuf==6.33.1
psutil==7.1.3
py-cpuinfo==9.0.0
pyclipper==1.3.0.post6
pycryptodome==3.23.0
pydantic==2.12.4
pydantic_core==2.41.5
pypdfium2==5.0.0
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytokens==0.3.0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.5
ruamel.yaml==0.18.16
ruamel.yaml.clib==0.2.15
safetensors==0.7.0
scikit-image==0.25.2
scipy==1.16.3
setuptools==80.9.0
shapely==2.1.2
shellescape==3.8.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sympy==1.14.0
tifffile==2025.10.16
torch==2.9.1
torchvision==0.24.1
tqdm==4.67.1
typer-slim==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
ujson==5.11.0
urllib3==2.5.0
wcwidth==0.2.14
